requirements.txt
pillow
numpy
opencv-python
pyautogui
nuitka